+++ ingestion/src/metadata/test_suite/api/workflow.py
@@ -46,7 +46,7 @@ from metadata.generated.schema.tests.testDefinition import TestDefinition
 from metadata.generated.schema.tests.testSuite import TestSuite
 from metadata.ingestion.api.parser import parse_workflow_config_gracefully
 from metadata.ingestion.api.processor import ProcessorStatus
-from metadata.ingestion.ometa.create_ometa import create_ometa_client
+from metadata.ingestion.ometa.client_utils import create_ometa_client
 from metadata.ingestion.ometa.ometa_api import OpenMetadata
 from metadata.interfaces.sqalchemy.sqa_test_suite_interface import SQATestSuiteInterface
 from metadata.orm_profiler.api.models import TablePartitionConfig
